Extract tests for isReferenceError to own file
